Immutize field - Cleanup code

+++ TestNinja/Mocking/EmployeeController.cs
@@ -4,7 +4,7 @@ namespace TestNinja.Mocking
 {
     public class EmployeeController
     {
-        private EmployeeContext _db;
+        private readonly EmployeeContext _db;
 
         public EmployeeController()
         {
@@ -14,8 +14,10 @@ namespace TestNinja.Mocking
         public ActionResult DeleteEmployee(int id)
         {
             var employee = _db.Employees.Find(id);
-            _db.Employees.Remove(employee);
+            if (employee != null)
+                _db.Employees.Remove(employee);
             _db.SaveChanges();
+
             return RedirectToAction("Employees");
         }
